atividades/webchat/fastapi/requirements.txt
fastapi
uvicorn[standard]
pyjwt
passlib[bcrypt]
bcrypt==3.2.2   # compatible with passlib
python-multipart